Getting API GET request to render on the page
