+++ Develop/.env.EXAMPLE
@@ -1,4 +1,3 @@
-node_modules
-.env
-.DS_Store
-willsFloatingReadME.md
+DB_NAME=db_name
+DB_USER=user
+DB_PASSWORD=password
