--- Develop/config/recipeAPI.js
@@ -1,1 +1,0 @@
-// Will probably contain configurations for connecting to and querying the recipe API.
