+++ Develop/package.json
@@ -6,7 +6,8 @@
     "scripts": {
         "test": "echo \"Error: no test specified\" && exit 1",
         "start": "node server.js",
-        "seed": "node seeds/seed.js"
+        "seed": "node seeds/seed.js",
+        "watch": "nodemon server.js"
     },
     "keywords": [],
     "author": "",
@@ -20,6 +21,7 @@
         "express-handlebars": "^5.2.0",
         "express-session": "^1.17.1",
         "mysql2": "^2.2.5",
+        "nodemon": "^2.0.22",
         "sequelize": "^6.3.5"
     },
     "devDependencies": {
